requirements.txt
python-periphery==2.4.1
